requirements.txt
streamlit>=1.30.0
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
google-generativeai>=0.7.0
requests>=2.31.0
pandas>=2.1.0
plotly>=5.18.0
fpdf2>=2.7.0
